requirements.txt
Flask==2.1.0
flask_swagger_ui==4.11.1
pymongo==4.3.3
Requests==2.31.0
Werkzeug==2.0.3
PyJWT==2.1.0
flask_cors==3.0.10
PyJWT==2.1.0
pika==1.3.2